+++ packages/dustkit/src/rpc/getMessagePortProvider.ts
@@ -1,16 +1,16 @@
 import { Provider, RpcRequest, RpcResponse, type RpcSchema } from "ox";
+import {
+  type CreateMessagePortOptions,
+  createMessagePort,
+} from "./createMessagePort";
 
-// TODO: add context that we send with each connection instead of init rpc method
 export function getMessagePortProvider<schema extends RpcSchema.Generic>({
   target,
   targetOrigin = "*",
-}: {
-  target: Window;
-  targetOrigin?: string;
-}): Provider.Provider<undefined, schema> {
-  const portPromise = createPort({ target, targetOrigin })
-    .catch(() => createPort({ target, targetOrigin }))
-    .catch(() => createPort({ target, targetOrigin }));
+}: CreateMessagePortOptions): Provider.Provider<undefined, schema> {
+  const portPromise = createMessagePort({ target, targetOrigin })
+    .catch(() => createMessagePort({ target, targetOrigin }))
+    .catch(() => createMessagePort({ target, targetOrigin }));
 
   const requestStore = RpcRequest.createStore<RpcSchema.Generic>();
   const provider = Provider.from({
@@ -37,43 +37,4 @@ export function getMessagePortProvider<schema extends RpcSchema.Generic>({
   });
 
   return provider as never;
-}
-
-export const initMessage = "MessagePortProvider";
-
-async function createPort({
-  target,
-  targetOrigin = "*",
-}: {
-  target: Window;
-  targetOrigin?: string;
-}) {
-  // TODO: configurable timeout
-  // TODO: retry with backoff
-  // TODO: add health check, recreate port if closed (see https://github.com/whatwg/html/issues/1766)
-
-  return new Promise<MessagePort>((resolve, reject) => {
-    const timeout = AbortSignal.timeout(1_000);
-    const channel = new MessageChannel();
-    channel.port1.addEventListener(
-      "message",
-      function onMessage(event) {
-        console.info("Got message from port", event);
-        if (event.data === "ready") {
-          resolve(channel.port1);
-        } else {
-          reject(new Error("Unexpected first message from MessagePort."));
-        }
-      },
-      { once: true },
-    );
-    channel.port1.start();
-    console.info("establishing MessagePortProvider with", targetOrigin);
-    target.postMessage(initMessage, targetOrigin, [channel.port2]);
-
-    timeout.addEventListener("abort", () => {
-      // TODO: clean up message channel/ports?
-      reject(timeout.reason);
-    });
-  });
 }
